+++ admin/src/components/Layout.tsx
@@ -10,6 +10,7 @@ import {
 } from "@headlessui/react";
 import { Bars3Icon, BellIcon, XMarkIcon } from "@heroicons/react/24/outline";
 import { Outlet, useLocation } from "react-router";
+import { Link } from "react-router-dom";
 
 function classNames(...classes: any[]) {
   return classes.filter(Boolean).join(" ");
@@ -17,11 +18,18 @@ function classNames(...classes: any[]) {
 
 export function Layout() {
   const location = useLocation();
-  const isSCIM = location.pathname.startsWith("/scim");
 
   const navigation = [
-    { name: "SAML Settings", href: "/", current: !isSCIM },
-    { name: "SCIM Settings", href: "/scim", current: isSCIM },
+    {
+      name: "SAML Settings",
+      href: "/saml",
+      current: location.pathname.startsWith("/saml"),
+    },
+    {
+      name: "SCIM Settings",
+      href: "/scim",
+      current: location.pathname.startsWith("/scim"),
+    },
   ];
   return (
     <>
@@ -38,9 +46,12 @@ export function Layout() {
           <div className="mx-auto max-w-7xl px-4 sm:px-6 lg:px-8">
             <div className="flex h-16 justify-between">
               <div className="flex">
-                <div className="flex flex-shrink-0 items-center text-sm">
+                <Link
+                  to="/"
+                  className="flex flex-shrink-0 items-center text-sm"
+                >
                   Settings Panel
-                </div>
+                </Link>
                 <div className="hidden sm:-my-px sm:ml-6 sm:flex sm:space-x-8">
                   {navigation.map((item) => (
                     <a
